[TFLite] NFC: Decouple pass manager setup from tf_to_tfl_flatbuffer This allows a separate binary to link with tf_to_tfl_flatbuffer and use it, but by providing its own pass manager.
PiperOrigin-RevId: 263806992
