+++ tensorflow/compiler/mlir/lite/python/BUILD
@@ -16,6 +16,7 @@ cc_library(
         "graphdef_to_tfl_flatbuffer.h",
     ],
     deps = [
+        "//tensorflow/compiler/mlir/lite:tf_tfl_passes",
         "//tensorflow/compiler/mlir/lite:tf_to_tfl_flatbuffer",
         "//tensorflow/compiler/mlir/tensorflow",
         "//tensorflow/compiler/mlir/tensorflow:convert_graphdef",
