+++ tensorflow/compiler/mlir/lite/python/graphdef_to_tfl_flatbuffer.cc
@@ -17,6 +17,7 @@ limitations under the License.
 
 #include "mlir/IR/MLIRContext.h"  // TF:local_config_mlir
 #include "mlir/IR/Module.h"  // TF:local_config_mlir
+#include "tensorflow/compiler/mlir/lite/tf_tfl_passes.h"
 #include "tensorflow/compiler/mlir/lite/tf_to_tfl_flatbuffer.h"
 #include "tensorflow/compiler/mlir/tensorflow/translate/import_model.h"
 #include "tensorflow/compiler/mlir/tensorflow/translate/mlir_roundtrip_flags.h"
@@ -137,10 +138,16 @@ Status ConvertGraphDefToTFLiteFlatBuffer(const toco::ModelFlags& model_flags,
   bool lower_tensor_list_ops = true;
   TF_ASSIGN_OR_RETURN(
       auto module, ConvertGraphdefToMlir(input, debug_info, specs, &context));
+
+  mlir::PassManager pm;
+  bool run_quantize = tensorflow::ShouldRunQuantizePasses(module.get());
+  tensorflow::AddTFToTFLConversionPasses(emit_builtin_tflite_ops, run_quantize,
+                                         emit_quant_adaptor_ops,
+                                         lower_tensor_list_ops, &pm);
   return ConvertTFExecutorToTFLOrFlatbuffer(
       module.get(), /*export_to_mlir=*/false, emit_builtin_tflite_ops,
       emit_select_tf_ops, emit_custom_ops, emit_quant_adaptor_ops,
-      lower_tensor_list_ops, result);
+      lower_tensor_list_ops, result, &pm);
 }
 
 }  // namespace tensorflow
