+++ tensorflow/compiler/mlir/lite/tf_to_tfl_flatbuffer.h
@@ -41,28 +41,6 @@ LoadFromGraphdefOrMlirSource(
     bool prune_unused_nodes, llvm::SourceMgr* source_mgr,
     mlir::MLIRContext* context);
 
-// Quantization passess will run only when the user specifies a quantized type
-// in the `-tf-inference-type` flag, which is converted to the function
-// attribute "tf.quantize" by the importer module.
-// TODO(fengliuai): switch to the cmd flag once the flags are moved to this
-// file with main method.
-bool ShouldRunQuantizePasses(mlir::ModuleOp m);
-
-// Add the MLIR passes that convert TF control flow dialect to TF Lite dialect
-// to a MLIR `pass_manager`. These passes first raise the control flow in the TF
-// control flow dialect, decode the constant tensors, and then legalize the
-// module to TF Lite dialect with some optimizations afterwards.
-// If `emit_builtin_tflite_ops` is true, TF Lite legalization passes will be
-// added, which produces TF Lite ops. If `run_quantize` is true, quantization
-// passes will be added. If `emit_quant_adaptor_ops` is true, Quantize and
-// Dequantize ops are added to the inputs and outputs of the quantized model.
-// If `lower_tensor_list_ops` is true, tensorlist ops will be lowered to basic
-// TF ops before legalization to TF Lite dialect.
-void AddTFToTFLConversionPasses(bool emit_builtin_tflite_ops, bool run_quantize,
-                                bool emit_quant_adaptor_ops,
-                                bool lower_tensor_list_ops,
-                                mlir::PassManager* pass_manager);
-
 // Taking a MLIR module in TF executor dialect and a set of parameters,
 // applies a set of passes to convert the module to TF Lite dialect and
 // serializes the result to a string. Depending on an attribute in the module
@@ -71,7 +49,8 @@ void AddTFToTFLConversionPasses(bool emit_builtin_tflite_ops, bool run_quantize,
 Status ConvertTFExecutorToTFLOrFlatbuffer(
     mlir::ModuleOp module, bool export_to_mlir, bool emit_builtin_tflite_ops,
     bool emit_select_tf_ops, bool emit_custom_ops, bool emit_quant_adaptor_ops,
-    bool lower_tensor_list_ops, std::string* result);
+    bool lower_tensor_list_ops, std::string* result,
+    mlir::PassManager* pass_manager);
 }  // namespace tensorflow
 
 #endif  // TENSORFLOW_COMPILER_MLIR_LITE_TF_TO_TFL_FLATBUFFER_H_
